# set metal fast math default to 0 (disabled)