It's a correctness fix because we use inf and nan. Let's see how slow it is