@geohot 2024-02-12
Huh, the failing test is strange. I've always thought there was some deep sync issue with METAL, maybe it's that.