@chenyuxyz 2024-02-12
note that speed comparison needs to be measured with no compiler cache