@chenyuxyz 2024-02-14
opened issue #3386 to track conv1d failure

i don't know why but it seems to be faster to run with fast math disabled on M1 Max
`DISABLE_COMPILER_CACHE=1 NOCLANG=1 python3 test/external/external_model_benchmark.py`

master
```
resnet50        tinygrad_metal_jitless      49.61 ms
resnet50        tinygrad_metal_jit          25.27 ms
resnet50        torch_cpu                   53.77 ms
resnet50        torch_mps                   20.65 ms
resnet50        onnxruntime_cpu             27.98 ms
resnet50        outputs validated with rtol=2.0e-03, atol=2.0e-03
openpilot       tinygrad_metal_jitless      62.43 ms
openpilot       tinygrad_metal_jit           6.02 ms
openpilot       torch_cpu                 1018.42 ms
openpilot       torch_mps                   26.32 ms
openpilot       onnxruntime_cpu             36.33 ms
openpilot       outputs validated with rtol=2.0e-03, atol=2.0e-03
efficientnet    tinygrad_metal_jitless      73.46 ms
efficientnet    tinygrad_metal_jit          13.22 ms
efficientnet    torch_cpu                  604.24 ms
efficientnet    torch_mps                   12.97 ms
efficientnet    onnxruntime_cpu             46.45 ms
efficientnet    outputs validated with rtol=2.0e-03, atol=2.0e-03
shufflenet      tinygrad_metal_jitless      29.44 ms
shufflenet      tinygrad_metal_jit           2.52 ms
shufflenet      torch_cpu                   26.35 ms
shufflenet      torch_mps                   62.14 ms
shufflenet      onnxruntime_cpu              4.17 ms
shufflenet      outputs validated with rtol=2.0e-03, atol=2.0e-03
commavq         tinygrad_metal_jitless     107.61 ms
commavq         tinygrad_metal_jit           9.43 ms
commavq         onnx2torch       NotImplementedError
commavq         onnxruntime_cpu             68.42 ms
commavq         outputs validated with rtol=2.0e-03, atol=2.0e-03
```

this pr
```
resnet50        tinygrad_metal_jitless      40.62 ms
resnet50        tinygrad_metal_jit          15.81 ms
resnet50        torch_cpu                   56.81 ms
resnet50        torch_mps                   22.50 ms
resnet50        onnxruntime_cpu             31.98 ms
resnet50        outputs validated with rtol=2.0e-03, atol=2.0e-03
openpilot       tinygrad_metal_jitless      64.82 ms
openpilot       tinygrad_metal_jit           5.77 ms
openpilot       torch_cpu                 1092.76 ms
openpilot       torch_mps                   27.42 ms
openpilot       onnxruntime_cpu             37.94 ms
openpilot       outputs validated with rtol=2.0e-03, atol=2.0e-03
efficientnet    tinygrad_metal_jitless      72.75 ms
efficientnet    tinygrad_metal_jit          12.13 ms
efficientnet    torch_cpu                  648.02 ms
efficientnet    torch_mps                   13.56 ms
efficientnet    onnxruntime_cpu             46.11 ms
efficientnet    outputs validated with rtol=2.0e-03, atol=2.0e-03
shufflenet      tinygrad_metal_jitless      30.97 ms
shufflenet      tinygrad_metal_jit           2.02 ms
shufflenet      torch_cpu                   30.68 ms
shufflenet      torch_mps                   63.37 ms
shufflenet      onnxruntime_cpu              4.21 ms
shufflenet      outputs validated with rtol=2.0e-03, atol=2.0e-03
commavq         tinygrad_metal_jitless     111.81 ms
commavq         tinygrad_metal_jit           9.39 ms
commavq         onnx2torch       NotImplementedError
commavq         onnxruntime_cpu             69.08 ms
commavq         outputs validated with rtol=2.0e-03, atol=2.0e-03
```