@chenyuxyz 2024-02-14
on M2 Max benchmark it seems to be ~2% faster (cifar) to ~2% slower (float tc gemm), probably within noise. given no significant impact i think it's worth it to merge this