@chenyuxyz 2024-02-14
with no significant perf impact, I think merging this to resolve all inf/nan problem on METAL is worth it. I am ready to merge this. @geohot let me know if you have any concerns or perf test suggestions